requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiorwlock==1.5.0
aiosignal==1.3.2
anyio==4.8.0
asqlite @ git+https://github.com/Rapptz/asqlite.git@8d929c7628e53d01b6905866f8b6d350e76a8d40
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.45.2
twitchio @ git+https://github.com/PythonistaGuild/TwitchIO.git@798da66f4eb8edcc6e94c9a26744e8edb52b3126
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3